+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/ ), and followin
 ### Changed
  - Cleaned up the messages sent to the user and logged when we are deleting/changing/creating chunk loaders.
 ### Fixed
+ - Fixed the log messages around edit/delete chunkloaders #37
 
 
 ## 2020-09-30 3.8.0